requirements.txt
groq>=0.34.0
datasets==2.14.0
pyarrow>=18.0.0
pandas>=2.2.2
numpy>=1.24.0,<2.0.0
matplotlib>=3.9.0
seaborn==0.13.0
huggingface_hub>=0.16.4,<0.18
transformers==4.35.0
torch>=2.6.0
tqdm==4.66.0
python-dotenv>=1.0.0
scikit-learn==1.3.0